% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_los.R
\name{aggregate_los}
\alias{aggregate_los}
\title{Aggregate duration of record level events}
\usage{
aggregate_los(
  start_date,
  end_date,
  level = diff(range(start_date, end_date)),
  strata = NULL,
  count_id = NULL,
  event_id = NULL,
  fill = FALSE,
  silent = TRUE,
  units = "days",
  los_range = NULL
)
}
\arguments{
\item{start_date}{Date. Event start date}

\item{end_date}{Date. Event end date}

\item{level}{Character. Desired level. Passed to 'by' in \code{\link{seq}}.
Options include all accepted in \code{\link[=seq.POSIXt]{seq.POSIXt(..., by = XX)}}}

\item{strata}{Atomic. Population strata}

\item{count_id}{Atomic. If provided the unique count of \code{count_id}
per iteration (batch) is included}

\item{event_id}{Atomic. If provided the unique count of \code{event_id}
per iteration (batch) is included}

\item{fill}{Logical. If TRUE, place holders are included if
there is no occurrence of the event within the batch of the loop.}

\item{silent}{Logical. Print (or not) current iteration of loop.}

\item{units}{Character. Desired unit for LOS. Passed to \code{\link{difftime}}.
Options include all accepted in \code{\link{difftime}}s}

\item{los_range}{NULL|Date. A date range of the period to cover in the aggregation.
If \code{NULL} (default), the rane of \code{start_date} and \code{start_date} is used.}
}
\value{
A data frame with the following columns
\itemize{
\item \code{strata} - See \code{strata} argument.
\item \code{days_origin_0} - Total length of stay per \code{strata} when the first day in each period was taken day 0.
\item \code{days_origin_1} - Total length of stay per \code{strata} when the first day in each period was taken day 1.
\item \code{unique.count_id} - Unique count of \code{count.id} per per \code{strata}.
\item \code{unique.event_id} - Unique count of \code{event_id} per per \code{strata}.
\item \code{batch} - Batch of length of stay per \code{strata} and \code{level}
\item \code{start_date} - Start date of \code{batch}
\item \code{end_date} - End date of \code{batch}
}
}
\description{
Aggregate duration of record level events
}
\examples{
# Example of dated events
dfr <- data.frame(
  start_date = as.Date(c('2024-05-28', '2024-05-05', '2024-06-01')),
  end_date = as.Date(c('2024-06-05', '2024-05-11', '2024-06-20')),
  strata = c('A', 'A', 'B'))

# Length of stay across the entire period
 dfr$duration <- difftime(dfr$end_date, dfr$start_date, 'days')
 dfr

 # Aggregate to a annual level
 aggregate_los(
   start_date = dfr$start_date,
   end_date = dfr$end_date,
   level = 'year',
   strata = dfr$strata
 )

 # Aggregate to a monthly level
 aggregate_los(
   start_date = dfr$start_date,
   end_date = dfr$end_date,
   level = 'month',
   strata = dfr$strata
 )

 # Cover a range of wider of dates. Use `fill` for empty rows
 aggregate_los(
   start_date = dfr$start_date,
   end_date = dfr$end_date,
   level = 'month',
   strata = dfr$strata,
   los_range =  c('2024-01-01', '2024-12-01')
 )

 aggregate_los(
   start_date = dfr$start_date,
   end_date = dfr$end_date,
   level = 'month',
   strata = dfr$strata,
   los_range =  c('2024-01-01', '2024-12-01'),
   fill = TRUE
 )

 # Return LOS is different units of time
 aggregate_los(
   start_date = dfr$start_date,
   end_date = dfr$end_date,
   level = 'month',
   strata = dfr$strata,
   los_range =  c('2024-01-01', '2024-12-01'),
   units = 'weeks'
 )


 # Return LOS is different units of time
 aggregate_los(
   start_date = dfr$start_date,
   end_date = dfr$end_date,
   level = '2 years',
   strata = dfr$strata,
   los_range =  c('2020-01-01', '2025-12-01'),
   units = 'weeks',
   fill = TRUE
 )

}
