% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_population_estimates.R
\name{aggregate_population_estimates}
\alias{aggregate_population_estimates}
\title{Aggregate population estimates}
\usage{
aggregate_population_estimates(estimate, period, strata, from, to)
}
\arguments{
\item{estimate}{Numeric. Population estimate}

\item{period}{Integer. Numeric codes for the corresponding period. See \link{period_cd}}

\item{strata}{Atomic. Population strata}

\item{from}{Character. Current level. "cy", "fy", "cq", "fq", "cm", "fm"}

\item{to}{Character. Desired level. "cy", "fy", "cq", "fq", "cm", "fm"}
}
\value{
A data frame
}
\description{
Aggregate population estimates
}
\examples{
dfr <- data.frame(
  ccg_cd = "E92001",
  cyear = c(2021, 2011),
  mid_yr_est = c(1457775, 1457783))


aggregate_population_estimates(
  strata = dfr$ccg_cd,
  estimate = dfr$mid_yr_est,
  period = dfr$cyear,
  from = "cy",
  to = "cq"
)

aggregate_population_estimates(
  strata = dfr$ccg_cd,
  estimate = dfr$mid_yr_est,
  period = dfr$cyear,
  from = "cy",
  to = "fq"
)

aggregate_population_estimates(
  strata = dfr$ccg_cd,
  estimate = dfr$mid_yr_est,
  period = dfr$cyear,
  from = "cy",
  to = "cm"
)

aggregate_population_estimates(
  strata = dfr$ccg_cd,
  estimate = dfr$mid_yr_est,
  period = dfr$cyear,
  from = "cy",
  to = "fm"
)
}
