% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_checks.R
\name{err_checks}
\alias{err_checks}
\alias{err_bys_rank_1}
\alias{err_bys_rank_2}
\title{Check for errors}
\usage{
err_bys_rank_1(..., by)

err_bys_rank_2(..., by)
}
\arguments{
\item{...}{...}
}
\description{
Check for errors
}
