% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt}
\alias{fmt}
\title{Common figure formats}
\usage{
fmt(x, fmt = "count", dp = 1)
}
\arguments{
\item{x}{Numerical vector to be formatted.}

\item{fmt}{Character vector specifying the type of format to be applied.}

\item{dp}{Decimal place}
}
\value{
A Character vector
}
\description{
Converts a number to favoured formats for inclusion in texts and reports.
}
\examples{

fmt(10.34456,"rate")

fmt(10.34456,"rate", dp=3)

fmt(.1,"rate", dp=3)

paste0(fmt(10.34456,"percent"),"\%")

fmt(1034456,"count")

fmt(.001,"rate")

fmt(.001,"p_val")

fmt(.001,"percent")


}
\seealso{
\code{\link{fy}}
}
