% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fy.R
\name{fy}
\alias{fy}
\title{Financial year labels as used in routine reports}
\usage{
fy(x, fmt = "abb")
}
\arguments{
\item{x}{Numerical vector - Financial year. Accepts; 2018, 201819 or 20182019.}

\item{fmt}{Character vector - "abb" or "full". Short hand format e.g. "2018/19" or full description e.g. "April 2018 to March 2019".}
}
\value{
Character - A description of the financial year.
}
\description{
Favoured formats for labelling financial years.
}
\examples{

fy(2018)

fy(201819)

fy(20182019)

}
\seealso{
\code{\link{fmt}}
}
