% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_funcs.R
\name{misc_funcs}
\alias{misc_funcs}
\alias{same_len_3dots}
\alias{len_3dots}
\title{Miscellaneous functions}
\usage{
same_len_3dots(...)

len_3dots(...)
}
\arguments{
\item{...}{...}
}
\description{
Miscellaneous functions
}
\details{
\bold{\code{same_len_3dots}} - Logical test for matching lengths in an \code{ellipsis (...)}

\bold{\code{len_3dots}} - Length of each entry in an \code{ellipsis (...)}
}
