% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_rates.R
\name{response_rates}
\alias{response_rates}
\title{Response rates}
\usage{
response_rates(x)
}
\arguments{
\item{x}{\code{list} of \code{atomic} vectors or \code{data.frame}}
}
\value{
\code{list}
}
\description{
Response rate for every value in a \code{list} of \code{atomic} vectors
}
\examples{
dataset <- data.frame(a = c("A", "C", "D"),
                      b = c("D", "A", "B"),
                      c = c("C", "E", "B"))

dataset
response_rates(dataset)
cbind(dataset, as.data.frame(response_rates(dataset)))
}
