% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_to.R
\name{round_to}
\alias{round_to}
\title{Round to multiples of a number.}
\usage{
round_to(x, to = 50)
}
\arguments{
\item{x}{Numeric. Figure to round down.}

\item{to}{Numeric. Multiple.}
}
\value{
Numeric
}
\description{
Round to multiples of a number.
}
\examples{

round_to(135, 10)
round_to(135, 2)

}
