% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_cd.R
\name{period_cd}
\alias{period_cd}
\alias{cm_to_fm}
\alias{fm_to_cm}
\alias{cq_to_fq}
\alias{fq_to_cq}
\title{Numeric codes for periods}
\usage{
period_cd(x, period = "fy")

cm_to_fm(f)

fm_to_cm(f)

cq_to_fq(f)

fq_to_cq(f)
}
\arguments{
\item{x}{\code{Date}, \code{POSIXct} or \code{POSIXt}.}

\item{f}{Time code. Output of \code{period_cd()}.}
}
\value{
\code{numeric}
}
\description{
Convenient time codes for analysis.
}
\details{
Convert calendar month codes to financial month codes

Convert financial month codes to calendar month codes

Convert calendar month codes to financial month codes

Convert financial month codes to calendar month codes
}
\examples{
dt <- Sys.Date()
period_cd(dt)
period_cd(dt,"cy")
period_cd(dt,"fq")
period_cd(dt,"cq")
period_cd(dt,"fm")
period_cd(dt,"cm")
period_cd(dt,"fd")
period_cd(dt,"cq")
}
