% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_to.R
\name{round_to}
\alias{round_to}
\title{Round to}
\usage{
round_to(x, to, f = round)
}
\arguments{
\item{val}{Ordered vector}

\item{by}{Increment}

\item{func}{Function}
}
\value{
Vector of he same class as \code{val}
}
\description{
Round to
}
\examples{

round_to(1104.3, 5)
round_to(1104.3, 5, ceiling)
round_to(1104.3, 5, trunc)
}
